Add iterative floorplanning algorithm
We use a grid of Pblocks (two per clock region row) as the floorplanning
granularity. This ensures 2 rows of cores can be placed in SLR2/SLR0, and
1 row of cores can be placed in SLR1 (since the shell removes half the width
of the row).

Places x0y0 at bottom-left of chip and xNyN at top-right of chip.
Then we rotate x0y0 until its assigned Pblock is the one where a user-provided
anchor is located.